Compiling working version of mstpcap on macOS 14 (Though should work on many earlier version,)
TODO: Needs cleanup. Add command line help to explain the usefull already built-in command-line flags
Figure out whether the needs to be seperate e.g. or can be integrated as an omnibus makefile that continues to handle the windows and linux version.
This is closest to bsd, but not quite, as it depends on an Apple only lib IOKit/serial/ioss.h need to fix makefile preprocessor stuff for per platform includes and code blocks if thats going to be the case.